perf(test): split cumulative suites across shards
